requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
deep-translator==1.11.4
spotipy==2.23.0
transformers==4.35.2
torch==2.2.2